Function/macro wrappers for memory management
This is necessary so we have access to the system malloc() and friends.
Normally, perl will muck about with them and then complain bitterly if
we use them for process-global memory management.

The macros exist for easy reimplementation with something less
expensive/silly than full C functions. "inline" was eschewed for
portability.