Use private memory management, not perl's
This is necessary to get the system malloc() et al instead of perl's.
Applies to both the XSAccessor guts and the hash table.

Alternative would be to use PerlMemShared_*, but that may need aTHX
occasionally.

(This was modified from the commit in the memoryCraziness branch to
include the memory leak fix that was introduced later in the branch.)