Segfault in very esoteric corner case (root cause perl5 bug)
I nailed it down, though I don't think this warrants any fixing on the XSA
part. The committed test fails on everything < 5.8.9. It appears to be caused
by this: http://rt.perl.org/rt3/Public/Bug/Display.html?id=35878

The only reason I proceeded writing a full test is that the *first* pass seems
to work, so It may be something relatively easy to fix in the guts of XSA.

If you are wondering why someone would be doing so much hackery when all we
want are quick accessors see lines 99~122:
http://dev.catalystframework.org/svnweb/bast/revision?rev=9746