Implement sort-of-Class::Accessor-style getter
The tagline isn't quite true. Basically, we offer the following in
optimized XS:

  sub get_foo {
    my $self = shift;
    if (not exists($self->{foo})) {
      $self->{foo} = $self->_get("foo");
    }
    return $self->{foo};
  }

Whereas Class::Accessor ALWAYS calls the "get()" method (note the lack
of an underscore there). So it's not a drop-in replacement and will
never be since calling into Perl from XS is super-expensive and defeats
the whole purpose of this excercise.

TODO: Consider whether this is a good idea to expose.
TODO: valgrind the f*cker